CatalogImpl: move validation rules to support class
